Change np.int to np.int16

+++ server/voice_changer/RVC/custom_vc_infer_pipeline.py
@@ -64,7 +64,7 @@ class VC(object):
         ) + 1
         f0_mel[f0_mel <= 1] = 1
         f0_mel[f0_mel > 255] = 255
-        f0_coarse = np.rint(f0_mel).astype(np.int)
+        f0_coarse = np.rint(f0_mel).astype(np.int16)
 
         # Volume Extract
         # volume = self.extractVolume(audio, 512)
